+++ tools/api-markdown-documenter/src/LintApiModel.ts
@@ -22,8 +22,8 @@ import {
 	DocNodeKind,
 } from "@microsoft/tsdoc";
 
-import type { ConfigurationBase } from "./ConfigurationBase.js";
 import { defaultConsoleLogger } from "./Logging.js";
+import type { LoggingConfiguration } from "./LoggingConfiguration.js";
 import { resolveSymbolicReference } from "./utilities/index.js";
 
 /**
@@ -31,7 +31,7 @@ import { resolveSymbolicReference } from "./utilities/index.js";
  *
  * @beta
  */
-export interface LintApiModelConfiguration extends ConfigurationBase {
+export interface LintApiModelConfiguration extends LoggingConfiguration {
 	/**
 	 * The API model to lint.
 	 */